core API: move Bitrate and PlayerConfig from core to playback
